Ignore empty fields properly
This will improve the overwall user experience as
previously you could click any field. If it was
empty, it wouldn't get highlighted, but still count.

When the user then clicked on a figure, he would
get an error that his alleged move was invalid
even though the user probably wanted to start one
and maybe just misclicked before.